+++ CHANGELOG.md
@@ -39,3 +39,4 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 * Fixed filtering issue with project overview.
 * Changed sprint report form to GET method.
 * Fixed worklog select path.
+* Optimized worklog select javascript.